feat: update dependencies and move code to src dir
