+++ package.json
@@ -6,10 +6,10 @@
     "node": ">=20.0 < 21.0"
   },
   "scripts": {
-    "start": "pnpm build && node dist/server.js",
-    "build": "pnpm fix && rimraf dist && tsc -p tsconfig.json",
-    "dev": "nodemon server.ts",
-    "lint": "eslint src",
+    "start": "pnpm build && node dist/index.js",
+    "build": "pnpm fix && rimraf dist && tsc",
+    "dev": "nodemon src/index.ts",
+    "lint": "eslint src/**/*",
     "format": "prettier src --check",
     "fix": "pnpm format --write && pnpm lint --fix"
   },
@@ -27,22 +27,22 @@
   "author": "Ahmad Azwar Annas <ahmadazw2@gmail.com>",
   "license": "GPL-3.0-or-later",
   "devDependencies": {
-    "@eslint/js": "^9.3.0",
-    "@types/node": "^20.12.12",
+    "@eslint/js": "^9.4.0",
+    "@types/node": "^20.14.2",
     "eslint": "~9.3.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import-x": "^0.5.1",
-    "nodemon": "^3.1.1",
-    "prettier": "^3.2.5",
+    "nodemon": "^3.1.3",
+    "prettier": "^3.3.1",
     "rimraf": "^5.0.7",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.11.0"
+    "typescript-eslint": "8.0.0-alpha.29"
   },
   "dependencies": {
     "@fastify/cors": "^9.0.1",
     "@fastify/swagger": "^8.14.0",
-    "@fastify/swagger-ui": "^3.0.0",
+    "@fastify/swagger-ui": "^3.1.0",
     "dotenv": "^16.4.5",
     "fastify": "^4.27.0",
     "puppeteer": "^22.10.0",
