+++ tsconfig.json
@@ -6,8 +6,8 @@
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
-    "outDir": "dist",
-    "rootDir": "."
+    "rootDir": "src",
+    "outDir": "dist"
   },
   "$schema": "https://json.schemastore.org/tsconfig"
 }
